requirements.txt
PyMySQL~=0.10.0
matplotlib~=3.3.1
numpy~=1.19.1
python-dotenv~=0.14.0
